refactor: Agendamentos components
